reformat toolbar
