+++ src/components/editor/editor-window/tool-bar/tool-bar.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 import { Button, ButtonToolbar } from 'react-bootstrap'
 import { ForkAwesomeIcon } from '../../../common/fork-awesome/fork-awesome-icon'
 import './tool-bar.scss'
+import { addMarkup, extractSelection, updateSelection } from './utils'
 
 export interface ToolBarProps {
   content: string
@@ -16,107 +17,12 @@ export const ToolBar: React.FC<ToolBarProps> = ({ content, startPosition, endPos
     alert('This feature is not yet implemented')
   }
 
-  const changeSelection = (selection: string) => {
-    const lines = content.split('\n')
-    const newLines = selection.split('\n')
-    const extraLines = newLines.length - 1
-    const selectionIncludeNewline = selection.includes('\n')
-    if (startPosition.line === endPosition.line && startPosition.ch === endPosition.ch) {
-      if (!selectionIncludeNewline) {
-        console.debug(lines[startPosition.line].slice(0, startPosition.ch) + selection + lines[startPosition.line].slice(startPosition.ch))
-        lines[startPosition.line] = lines[startPosition.line].slice(0, startPosition.ch) + selection + lines[startPosition.line].slice(startPosition.ch)
-      } else {
-        for (let i = lines.length - 1; i > startPosition.line; i--) {
-          lines[i + extraLines] = lines[i]
-        }
-        const lastPart = lines[startPosition.line].slice(startPosition.ch)
-        lines[startPosition.line] = lines[startPosition.line].slice(0, startPosition.ch) + newLines[0]
-        let j = 1
-        for (let i = 0; i < extraLines; i++) {
-          lines[i + startPosition.line + 1] = newLines[j]
-          j++
-        }
-        lines[extraLines + startPosition.line] += lastPart
-      }
-    }
-    if (startPosition.line === endPosition.line && startPosition.ch !== endPosition.ch) {
-      if (!selectionIncludeNewline) {
-        lines[startPosition.line] = lines[startPosition.line].slice(0, startPosition.ch) + selection + lines[startPosition.line].slice(endPosition.ch)
-      } else {
-        for (let i = lines.length - 1; i > startPosition.line; i--) {
-          lines[i + extraLines] = lines[i]
-        }
-        const lastPart = lines[startPosition.line].slice(endPosition.ch)
-        lines[startPosition.line] = lines[startPosition.line].slice(0, startPosition.ch) + newLines[0]
-        let j = 1
-        for (let i = 0; i < extraLines; i++) {
-          lines[i + startPosition.line + 1] = newLines[j]
-          j++
-        }
-        lines[extraLines + startPosition.line] += lastPart
-      }
-    }
-    if (startPosition.line !== endPosition.line && startPosition.ch !== endPosition.ch) {
-      console.debug('not the same')
-      console.debug('before', lines)
-      for (let i = lines.length - 1; i > endPosition.line; i--) {
-        lines[i + extraLines] = lines[i]
-      }
-      let j = 0
-      for (let i = 0; i <= extraLines; i++) {
-        lines[i + startPosition.line] = newLines[j]
-        j++
-      }
-      console.debug('after', lines)
-    }
-    onContentChange(lines.join('\n'))
-  }
+  const changeSelection = (selection: string) => updateSelection(content, startPosition, endPosition, onContentChange, selection)
+  const getSelection = () => extractSelection(content, startPosition, endPosition)
 
-  const getSelection = (): string => {
-    if (startPosition.line === endPosition.line && startPosition.ch === endPosition.ch) {
-      return ''
-    }
-
-    const lines = content.split('\n')
-
-    if (startPosition.line === endPosition.line) {
-      return lines[startPosition.line].slice(startPosition.ch, endPosition.ch)
-    }
-
-    let newSelection = lines[startPosition.line].slice(startPosition.ch) + '\n'
-    for (let i = startPosition.line; i <= endPosition.line; i++) {
-      if (i === endPosition.line) {
-        newSelection += lines[i].slice(0, endPosition.ch) + '\n'
-      } else {
-        newSelection += lines[i] + '\n'
-      }
-    }
-    return newSelection
-  }
-
-  const makeSelectionBold = () => {
-    const selection = getSelection()
-    if (selection === '') {
-      return
-    }
-    changeSelection(`**${selection}**`)
-  }
-
-  const makeSelectionItalic = () => {
-    const selection = getSelection()
-    if (selection === '') {
-      return
-    }
-    changeSelection(`*${selection}*`)
-  }
-
-  const strikeThroughSelection = () => {
-    const selection = getSelection()
-    if (selection === '') {
-      return
-    }
-    changeSelection(`~~${selection}~~`)
-  }
+  const makeSelectionBold = () => addMarkup(content, startPosition, endPosition, onContentChange, '**')
+  const makeSelectionItalic = () => addMarkup(content, startPosition, endPosition, onContentChange, '*')
+  const strikeThroughSelection = () => addMarkup(content, startPosition, endPosition, onContentChange, '~~')
 
   const addHeaderLevel = () => {
     const lines = content.split('\n')
@@ -211,7 +117,7 @@ export const ToolBar: React.FC<ToolBarProps> = ({ content, startPosition, endPos
         <ForkAwesomeIcon icon="link"/>
       </Button>
       <Button onClick={notImplemented}>
-        <ForkAwesomeIcon icon="file-image-o"/>
+        <ForkAwesomeIcon icon="picture-o"/>
       </Button>
       <Button onClick={notImplemented}>
         <ForkAwesomeIcon icon="upload"/>
